+++ app/src/main/java/nz/ac/auckland/lablet/camera/VideoImportPlugin.java
@@ -42,7 +42,7 @@ public class VideoImportPlugin extends AbstractFileImportPlugin {
         try {
             StorageLib.copyFile(importFile, new File(dataStorageDir, fileName), new StreamHelper.IProgressListener() {
                 @Override
-                public void onNewProgress(int totalProgress) {
+                public void onNewProgress(long totalProgress) {
 
                 }
             }, 32 * 1024);
